Move log setup methods to utils functions.
Move setup rabbitmq handler and log file handler to mash utils
module. These can be shared with API based services.
