+++ mash/utils/mash_utils.py
@@ -18,6 +18,7 @@
 
 import datetime
 import json
+import logging
 import os
 import random
 import requests
@@ -30,7 +31,8 @@ from contextlib import contextmanager, suppress
 from string import ascii_lowercase
 from tempfile import NamedTemporaryFile
 
-from mash.mash_exceptions import MashException
+from mash.log.handler import RabbitMQHandler
+from mash.mash_exceptions import MashException, MashLogSetupException
 from mash.utils.json_format import JsonFormat
 
 
@@ -175,3 +177,42 @@ def handle_request(url, endpoint, method, job_data=None):
         raise MashException(msg)
 
     return response
+
+
+def setup_logfile(logfile):
+    """
+    Create log dir and log file if either does not already exist.
+    """
+    try:
+        log_dir = os.path.dirname(logfile)
+        if not os.path.isdir(log_dir):
+            os.makedirs(log_dir)
+    except Exception as e:
+        raise MashLogSetupException(
+            'Log setup failed: {0}'.format(e)
+        )
+
+    logfile_handler = logging.FileHandler(
+        filename=logfile, encoding='utf-8'
+    )
+
+    return logfile_handler
+
+
+def get_logging_formatter():
+    return logging.Formatter(
+        '%(newline)s%(levelname)s %(asctime)s %(name)s%(newline)s'
+        '    %(job)s %(message)s'
+    )
+
+
+def setup_rabbitmq_log_handler(host, username, password):
+    rabbit_handler = RabbitMQHandler(
+        host=host,
+        username=username,
+        password=password,
+        routing_key='mash.logger'
+    )
+    rabbit_handler.setFormatter(get_logging_formatter())
+
+    return rabbit_handler
